[ES|QL] Cleanups the operators generation  (#222772)
## Summary

Closes https://github.com/elastic/kibana/issues/218358

- Cleanups the operators hardcoded signatures. There were cases that I
could not avoid to add some extra signatures but there are been added to
the ones coming from ES so it is ok
- Organizes this script file and makes it more cleaner and reasable. It
has become impossible to navigate so I hope this change improves things

### Checklist

- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit d6b359d67a189872bd27eddbbab88f646985dc5c)
